requirements.txt
beautifulsoup4==4.13.3
python-dotenv==1.1.0
selenium==4.31.0
selenium-wire
Pillow
llama_parse